+++ tests/Unit/CurlTest.php
@@ -17,6 +17,7 @@ namespace CrowdSec\LapiClient\Tests\Unit;
 
 use CrowdSec\LapiClient\Bouncer;
 use CrowdSec\LapiClient\Tests\MockedData;
+use CrowdSec\LapiClient\TimeoutException;
 
 /**
  * @uses \CrowdSec\LapiClient\Configuration::getConfigTreeBuilder
@@ -30,6 +31,8 @@ use CrowdSec\LapiClient\Tests\MockedData;
  * @covers \CrowdSec\LapiClient\Bouncer::getStreamDecisions
  * @covers \CrowdSec\LapiClient\Bouncer::getFilteredDecisions
  * @covers \CrowdSec\LapiClient\Bouncer::manageRequest
+ * @covers \CrowdSec\LapiClient\Bouncer::getAppSecDecision
+ * @covers \CrowdSec\LapiClient\Bouncer::manageAppSecRequest
  */
 final class CurlTest extends AbstractClient
 {
@@ -70,6 +73,105 @@ final class CurlTest extends AbstractClient
             json_decode(MockedData::DECISIONS_FILTER, true),
             $decisionsResponse,
             'Success get filtered decisions'
+        );
+    }
+
+    public function testAppSecDecision()
+    {
+        // Success test
+        $mockCurlRequest = $this->getCurlMock(['exec', 'getResponseHttpCode']);
+        $mockCurlRequest->method('exec')->willReturn(
+            MockedData::APPSEC_ALLOWED
+        );
+        $mockCurlRequest->method('getResponseHttpCode')->willReturn(
+            MockedData::HTTP_200
+        );
+        $client = new Bouncer($this->configs, $mockCurlRequest);
+        $headers = [
+            'X-Crowdsec-Appsec-Ip' => 'test-value',
+            'X-Crowdsec-Appsec-Host' => 'test-value',
+            'X-Crowdsec-Appsec-User-Agent' => 'test-value',
+            'X-Crowdsec-Appsec-Verb' => 'test-value',
+            'X-Crowdsec-Appsec-Uri' => 'test-value',
+            'X-Crowdsec-Appsec-Api-Key' => 'test-value',
+        ];
+        $appSecResponse = $client->getAppSecDecision('GET', $headers);
+
+        $this->assertEquals(
+            json_decode(MockedData::APPSEC_ALLOWED, true),
+            $appSecResponse,
+            'Success get appsec decision'
+        );
+    }
+
+    public function testAppSecDecisionWithTimeout()
+    {
+        // Success test
+        $mockCurlRequest = $this->getCurlMock(['exec', 'error', 'errno']);
+        $mockCurlRequest->method('exec')->willReturn(false);
+        $mockCurlRequest->method('errno')->willReturn(\CURLE_OPERATION_TIMEOUTED);
+        $mockCurlRequest->method('error')->willReturn('Operation timed out');
+
+        $client = new Bouncer($this->configs, $mockCurlRequest);
+        $headers = [
+            'X-Crowdsec-Appsec-Ip' => 'test-value',
+            'X-Crowdsec-Appsec-Host' => 'test-value',
+            'X-Crowdsec-Appsec-User-Agent' => 'test-value',
+            'X-Crowdsec-Appsec-Verb' => 'test-value',
+            'X-Crowdsec-Appsec-Uri' => 'test-value',
+            'X-Crowdsec-Appsec-Api-Key' => 'test-value',
+        ];
+
+        $error = false;
+        $message = '';
+        try {
+            $client->getAppSecDecision('GET', $headers);
+        } catch (TimeoutException $e) {
+            $error = true;
+            $message = $e->getMessage();
+        }
+
+        $this->assertEquals(
+            true,
+            $error,
+            'A timeout should be thrown'
+        );
+
+        $this->assertEquals(
+            'CURL call timeout: Operation timed out',
+            $message,
+            'A timeout should be thrown'
+        );
+    }
+
+    public function testFilteredDecisionsWithTimeout()
+    {
+        // Success test
+        $mockCurlRequest = $this->getCurlMock(['exec', 'error', 'errno']);
+        $mockCurlRequest->method('exec')->willReturn(false);
+        $mockCurlRequest->method('errno')->willReturn(\CURLE_OPERATION_TIMEOUTED);
+        $mockCurlRequest->method('error')->willReturn('Operation timed out');
+
+        $client = new Bouncer($this->configs, $mockCurlRequest);
+
+        $error = false;
+        $message = '';
+        try {
+            $client->getFilteredDecisions();
+        } catch (TimeoutException $e) {
+            $error = true;
+            $message = $e->getMessage();
+        }
+        $this->assertEquals(
+            true,
+            $error,
+            'A timeout should be thrown'
+        );
+
+        $this->assertEquals(
+            'CURL call timeout: Operation timed out',
+            $message,
+            'A timeout should be thrown'
         );
     }
 }